refactor

+++ utilities/lavagna.h
@@ -123,12 +123,13 @@ card_t* lavagna_trova_card_per_id(int id);
 
 
 /**
+ * @todo Compattare tutto in un'unica funzione
  * @brief funzione che permette di popolare un vettore di uint16_t con le porte
  * degli utenti attualmente registrati nel sistema.
  * 
  * @param buf il buffer all'interno del quale inserire la lista delle porte utente.
  */
 
-int lavagna_user_list_to_vec(int* buf);
+int lavagna_user_list_to_vec(uint16_t* buf);
 
 #endif /* LAVAGNA_H */
